+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* [#49](https://github.com/dduugg/yard-sorbet/issues/49) Apply `@abstract` tags to `abstact!`/`interface!` modules
 * [#43](https://github.com/dduugg/yard-sorbet/issues/43) Add `T::Enum` support
 
 ### Bug fixes